requirements/dev.txt
# wheel>=0.34.2                   # Wheel
# Django==3.0.7                   # Django package
# pillow==7.1.0                   # Image manipulation
# djangorestframework==3.10.3     # DRF framework
django-dbbackup==3.3.0          # Database backup / restore functionality
django-cors-headers==3.2.0      # CORS headers extension for DRF
# django_filter==2.2.0            # Extended filtering options
django-mptt==0.11.0             # Modified Preorder Tree Traversal
django-sql-utils==0.5.0         # Advanced query annotation / aggregation
django-markdownx==3.0.1         # Markdown form fields
django-markdownify==0.8.0       # Markdown rendering
coreapi==2.3.0                  # API documentation
pygments==2.2.0                 # Syntax highlighting
# tablib==0.13.0                  # Import / export data files
django-crispy-forms==1.8.1      # Form helpers
django-import-export==2.0.0     # Data import / export for admin interface
django-cleanup==5.1.0           # Manage deletion of old / unused uploaded files
flake8==3.8.3                   # PEP checking
pep8-naming==0.11.1             # PEP naming convention extension
coverage==5.3                   # Unit test coverage
coveralls==2.1.2                # Coveralls linking (for Travis)
rapidfuzz==0.7.6                # Fuzzy string matching
django-stdimage==5.1.1          # Advanced ImageField management
django-weasyprint==1.0.1        # HTML PDF export
django-debug-toolbar==2.2       # Debug / profiling toolbar
django-admin-shell==0.1.2       # Python shell for the admin interface
django-money==1.1               # Django app for currency management
certifi                         # Certifi is (most likely) installed through one of the requirements above
django-error-report==0.2.0      # Error report viewer for the admin interface
django-test-migrations==1.1.0   # Unit testing for database migrations 
# python-barcode[images]==0.13.1  # Barcode generator
# qrcode[pil]==6.1                # QR code generator

# inventree                       # Install the latest version of the InvenTree API python library
